Support Git rev-parse
This adds support to get branch via git's rev-parse when branch is not
provided. Prevously we would immediately default to "main" when branch
wasn't provided however there are some obvious issues with this

1. While there is a recent push to move to "main" as a default branch
"master" is still reguarly used.
2. Not all projects use either "main" or "master" as the default branch.

Given we are doing a git clone there really isnt a need to just default
to some constant and instead we can determine the appropriate branch via
`rev-parse`. If for some reason `rev-parse` fails we then default to
"main". It might be better to fail instead of default but for now I
think defaulting is fine and we can always change that behavior.

One outstanding TODO related to this work is to hoist logic for
determining branch/source link higher in the call stack. The primary
reason is that we are executing `rev-parse` for each source file which
is unnecessary work.