Dont override source_links_enabled
We want to allow users to specify source_links_enabled on specific target snippets to disable include snippet source links. This would allow them to have a default template that allows source links to be added and override behavior for specific snippets when/if requires. Note that if users dont want to include source links they can use a template that doesnt include it.

This commit changes the behavior to only set `source_links_enabled` to true if `source_links_enabled` has not been provided and a source link was created.