Support URL and File Target Snippet
Previously snippet identifiers in target files must have been
associated with a snippet that was extracted from a source. With this
commit we allow snippets in target files to URL as well as full files.

If a URL is specified snippext will download the contents from the
specified URL and insert the entire content within the begin/end snippet
comment.

If snippet id cannot be found in the extract snippets, and is not a URL,
we'll assume its a file path. Snippext will attempt to read the file and
insert the entire content of hte file within the being/end snippet
comment. Currently we dont verify its a source file, but probably
should. We'll likely include that when we do some future refactoring.

Including in this commit is a change to how source links are buit.
Previously source links were built deep in the stack at the time of
rendering. One issue with this was that in order to determine appropriate git
branch when not specified required us to run `git::abbrev_ref` on every
snippet. In this commit we were able to hoist source link logic up into
the extract phase. While there is still some improvements that need to
be made within the make extract logic this s an overall improvement.

Closes #8