Extraction and Source Link Updates
This commit contains two primary items

1. Changes how snippets are extract - Previously we had two distinct
phases. First we would get all the source files for a specific source
and then we would iterate and extract. Now We extract snippets as we
read source file. The primary motivation was we and to keep details
about the source files so that we could pass the details when we
extracted snippets in order to build source links. Now that we extract
as we go we can build source links and attach it to the snippet without
needing to preservce details about the source.

Which brings us to the second item

2. Change how we build source links - Previouly we had source link logic
is a few places. Now we added a SourceLink type that tries to colocate
logic around building source links in a mostly common spot, building
source links for git is still a bit complicated, but in general there
is a bit more cohesion around source links or at least I hope