Add Omit Source Links Configuration and CLI Args
Previously we only allowed users to enable/disable source links on a per
snippet basis via `source_links_enabled`. With this commit we allow
users to omit source links via Snippext a `omit_source_link` configuration field
as well as via a `extract` cli flag `omit_source_link`.

We wanted the default behavior be to have snippets included so we
decided to replace `source_links_enabled` and align around
`omit_source_links` as the global configuration and `omit_source_link`
as the per snippet configuration. This makes it easy to have the omit
source link(s) booleans default to false. Given that we were flipping
the value of true from display source links to true meaning omitting we
needed to update the handlebar template from using `if` to using
`unless`. This caused us to change the data that was added to the
template rendering context. Handlebar's `unless` helper will render content
if the expression returns a falsy value. Falsy value in this context
means undefined, null, "", or false. This caused an issue because our
template context has Map<String, String> and "false" is not a falsy
value which means that `#unless "false"` would cause the source
link to not be included. In order to pass in a bool rather than a String
we changed the context to Map<String, serde_json::Value>

closes #13