Output Snippet For Each Template Defined
Previously we would only output the `default` template when rendering
snippets to the output directory. With this commit we will output a file
for each template. For example, If we have a snippet with an identifier of
`example_snippet` and we have two templates `default` and `raw` this  will
generate two files to the output directory assuming the output extension is `md`

- example_snippet_default.md
- example_snippet_raw.md

Closes #14