Allow Users to Retain Nested Snippet Comments
Previously when there was a nested snippet when the parent snippet
would render the nested snippet comments would be removed. For example
if we had the following source

```
// snippet::start main
fn main() {

    // snippet::start nested
    println!("printing...")
    // snippet::end
}
// snippet::end
```

The content for the `main` snippet would be

```
fn main() {

    println!("printing...")
}
```

With this commit we allow users to provide
`retain_nested_snippet_comments` configuration flag to keep the nested
snippet comments causing the above source snippet content to be

```
fn main() {

    // snippet::start nested
    println!("printing...")
    // snippet::end
}
```

Which shows that the `nested` snippet comments are preserved.

The `retain_nested_snippet_comments` flag can be provided in the
following ways
1. Via `retain_nested_snippet_comments` flag on the `extract` CLI
command. This will globally retain nested snippet comments.
2. Via `retain_nested_snippet_comments` option within the snippext
config file. This will globally retain nested snippet comments.
3. Via `retain_nested_snippet_comments` attribute on the source snippet.
This is a per-snippet configuration which will result in snippet
comments to be retained only for the specific snippet.

Closes #16