Add Example of Custom Snippet Attributes
This commit provides a test to verify users can provide custom snippet
attributes and use them within target snippets. The example provided in
the test is to provide `highlighted_lines` which is then used in the
target snippet to set `hl_lines`. This is an example of  how a snipet
would be used in MKDocs to have it highlight lines in the source block

Closes #12