Autodetect Source Language
Previously users had to explicitely include a `lang` attribute in the
source snippet if they wanted to have it included it when the snippet is
rendered.

With this commit we added `hyperpolygot`, a language detector based on
the popular ruby gem `linguist`, which is used to determine the language of
source files which we use to set the `lang` attribute on all snippets
within the source file. We also included an option to allow users to turn of
autodetection. Users can pass in a `disable_language_autodetect` flag on
 the `extract` command or set `enable_autodetect_language` in the
 snippet configuration file.

 Closes #17